fix: wrap update password in suspense
